clean up casting

+++ src/motion/mot_walk.c
@@ -329,8 +329,8 @@ void mot_walk(void) {
 f32 marioGetDashSpd(void) {
     MarioWork* mario;
     f32 dashSpeed;
-    u8 stickDir2;
-    u8 stickDir1;
+    s8 stickDir2;
+    s8 stickDir1;
 
     mario = marioGetPtr();
     dashSpeed = mario->mBaseDashSpeed;
@@ -339,7 +339,7 @@ f32 marioGetDashSpd(void) {
         stickDir2 = mario->wStickDir2;
         dashSpeed = mario->mBaseWalkSpeed;
 
-        if (((s8)stickDir1 * (s8)stickDir1) + ((s8)stickDir2 * (s8)stickDir2) <= 0xBD1) {
+        if ((stickDir1 * stickDir1) + (stickDir2 * stickDir2) <= 0xBD1) {
             dashSpeed *= 0.5f;
         }
     } else if (marioBgmodeChk() == 1) {
@@ -352,8 +352,8 @@ f32 marioGetDashSpd(void) {
 
 f32 marioGetWalkSpd(void) {
     f32 walkSpeed;
-    u8 stickDir2;
-    u8 stickDir1;
+    s8 stickDir2;
+    s8 stickDir1;
     MarioWork* mario;
 
     mario = marioGetPtr();
@@ -361,7 +361,7 @@ f32 marioGetWalkSpd(void) {
     if (mario->flags & MARIO_FLAG_PAPER_MODE) {
         stickDir1 = mario->wStickDir1;
         stickDir2 = mario->wStickDir2;
-        if (((s8) stickDir1 * (s8) stickDir1) + ((s8) stickDir2 * (s8) stickDir2) <= 0xBD1) {
+        if ((stickDir1 * stickDir1) + (stickDir2 * stickDir2) <= 0xBD1) {
             walkSpeed *= 0.5f;
         }
     } else if (marioBgmodeChk() == 1) {
